Modify "skillLevel" type - from "int" to "short"

+++ core/src/com/gpteam/shopmanager/Skill/Skills/Expertise.java
@@ -8,16 +8,16 @@ import com.gpteam.shopmanager.Variables.Variables;
  * Created by masmix on 20.01.2017.
  */
 public class Expertise {
-    private int skillLevel;
+    private short skillLevel;
 
-    public Expertise(int skillLevel) {
+    public Expertise(short skillLevel) {
         if (skillLevel > Variables.MAX_SKILL_LEVEL || skillLevel < Variables.MIN_SKILL_LEVEL)
             throw new IllegalArgumentException("Wrong skill amount (0 - 120). Found: " + skillLevel + ".");
         else
             this.skillLevel = skillLevel;
     }
 
-    public int getSkillLevel() {
+    public short getSkillLevel() {
         return skillLevel;
     }
